refactor ajax requests, create a food events file to handle event listeners, separated logic into proper files
