+++ lib/index.js
@@ -1,3 +1,3 @@
 require('./styles/styles.css');
 require('./ajax_requests/meals.js');
-require('./ajax_requests/foods.js')
+require('./event_listeners/food_events.js')
